refactor(architecture): reorganize project structure for clean architecture
