+++ src/application/modules/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { PrismaModule } from './prisma.module';
+import { PrismaModule } from '../../../infrastructure/database/prisma.module';
 
 @Module({
   imports: [PrismaModule],
