+++ src/application/modules/todos/todos.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
+import { PrismaService } from '../../../infrastructure/database/prisma.service';
 
 @Injectable()
 export class TodosService {
